Fix fetch item
