+++ modules/fetchLocal.js
@@ -1,0 +1,11 @@
+export default () => {
+  let BookList;
+  if (localStorage.getItem('BookList')) {
+    const toParse = JSON.parse(localStorage.getItem('BookList'));
+    BookList = toParse;
+  } else {
+    BookList = [];
+  }
+
+  return BookList;
+};
